a big update

+++ controllers/user.js
@@ -4,7 +4,6 @@ const Order = require('../models/order')
 const bcrypt = require('bcrypt');
 
 
-
 exports.add = (req, res, next) => {
     const name = req.body.name
     const username = req.body.username
@@ -33,7 +32,7 @@ exports.add = (req, res, next) => {
             return bcrypt.hash(req.body.password , 12)
         }
     }).then(result =>{
-        user = new User({name: name, username: username, email: email, password: result, imageUrl: imageUrl, isAdmin: false, cart: []})
+        user = new User({name: name, username: username, email: email, password: result, imageUrl: imageUrl, isAdmin: false, cart: { productList: [], total: 0}})
         return user.save()
     }).then(result => {
         res.status(201).json({ message: "User created!", user: user});
@@ -46,6 +45,12 @@ exports.add = (req, res, next) => {
 };
 
 exports.edit = (req, res, next) => {
+    if(req.userId != req.body._id){
+        let err = new Error()
+        err.statusCode=422
+        err.message="Invalid Token!"
+        throw err
+    }
     const _id = req.body._id
     const name = req.body.name
     const username = req.body.username
@@ -108,6 +113,12 @@ exports.edit = (req, res, next) => {
 };
 
 exports.delete = (req, res, next) => {
+    if(req.userId != req.body._id){
+        let err = new Error()
+        err.statusCode=422
+        err.message="Invalid Token!"
+        throw err
+    }
     const _id = req.body._id
     User.findByIdAndRemove(_id)
     .then(result =>{
@@ -129,6 +140,12 @@ exports.delete = (req, res, next) => {
 };
 
 exports.get = (req, res, next) => {
+    if(req.userId != req.body._id){
+        let err = new Error()
+        err.statusCode=422
+        err.message="Invalid Token!"
+        throw err
+    }
     const _id = req.body._id
     User.findById(_id)
     .then(result =>{
@@ -149,10 +166,45 @@ exports.get = (req, res, next) => {
     })
 };
 
+exports.adminGetUser = (req, res, next) => {
+    if(isAdmin){
+        const _id = req.body._id
+        User.findById(_id)
+        .then(result =>{
+            if(result != null){
+                res.status(200).json({ message: "User found!", user: result});
+            }
+            else{
+                let err = new Error()
+                err.statusCode=404
+                err.message="User not found!"
+                throw err
+            }
+        }).catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+              }
+              next(err);
+        })
+    }
+    else{
+        let err = new Error()
+        err.statusCode=422
+        err.message="Not authenticated!"
+        next(err)
+    }
+};
+
 exports.addToCart = (req, res, next) => {
+    if(req.userId != req.body.userId){
+        let err = new Error()
+        err.statusCode=422
+        err.message="Invalid Token!"
+        throw err
+    }
     const userId = req.body.userId
     const productId = req.body.productId
-    const quantity = req.body.quantity || 1
+    const quantity = +req.body.quantity || 1
     let user
     User.findById(userId)
     .then(result =>{
@@ -167,18 +219,20 @@ exports.addToCart = (req, res, next) => {
             throw err
         }
     }).then(result =>{
+        const totalDiffrence = quantity * result.price
         if(result != null){
             let productFound = false
-            for(let i= 0; i< user.cart.length; i++){
-                if(user.cart[i].product == productId){
-                    if(+user.cart[i].quantity + +quantity > result.quantity){
+            for(let i= 0; i< user.cart.productList.length; i++){
+                if(user.cart.productList[i].product == productId){
+                    if(+user.cart.productList[i].quantity + +quantity > result.quantity){
                         let err = new Error()
                         err.statusCode=422
                         err.message="Requierd quantity is not avilable!"
                         err.data = {limit:result.quantity}
                         throw err
                     }
-                    user.cart[i].quantity = +user.cart[i].quantity + +quantity
+                    user.cart.productList[i].quantity = +user.cart.productList[i].quantity + +quantity
+                    user.cart.total = user.cart.total + totalDiffrence
                     productFound = true 
                 }
             }
@@ -190,7 +244,8 @@ exports.addToCart = (req, res, next) => {
                     err.data = {limit:result.quantity}
                     throw err
                 }
-                user.cart.push({product: productId, quantity: +quantity})
+                user.cart.productList.push({product: productId, quantity: +quantity})
+                user.cart.total = user.cart.total + totalDiffrence
             }
             return user.save()
         }
@@ -211,9 +266,15 @@ exports.addToCart = (req, res, next) => {
 };
 
 exports.removeFromCart = (req, res, next) => {
+    if(req.userId != req.body.userId){
+        let err = new Error()
+        err.statusCode=422
+        err.message="Invalid Token!"
+        throw err
+    }
     const userId = req.body.userId
     const productId = req.body.productId
-    const quantity = req.body.quantity || 1
+    const quantity = +req.body.quantity || 1
     let user
     User.findById(userId)
     .then(result =>{
@@ -229,22 +290,28 @@ exports.removeFromCart = (req, res, next) => {
         }
     }).then(result =>{
         if(result != null){
+            let totalDiffrence = result.price
             let productFound = false
-            for(let i= 0; i< user.cart.length; i++){
-                if(user.cart[i].product == productId){
+            for(let i= 0; i< user.cart.productList.length; i++){
+                if(user.cart.productList[i].product == productId){
                     if(quantity == 0){
-                        user.cart[i].quantity = +quantity
+                        totalDiffrence = totalDiffrence * user.cart.productList[i].quantity
+                        user.cart.productList[i].quantity = 0
                     }
                     else{
-                        user.cart[i].quantity = +user.cart[i].quantity - +quantity
+                        totalDiffrence = totalDiffrence * quantity
+                        user.cart.productList[i].quantity = +user.cart.productList[i].quantity - +quantity
                     }
-                    if(user.cart[i].quantity < 1){
-                        const newCart = user.cart.filter(cartItem => {
+                    if(user.cart.productList[i].quantity < 1){
+                        if(user.cart.productList[i].quantity < 0){
+                            totalDiffrence = totalDiffrence + (user.cart.productList[i].quantity * result.price)
+                        }
+                        const newCart = user.cart.productList.filter(cartItem => {
                             if(cartItem.product != productId){
                                 return cartItem
                             }
                         })
-                        user.cart = newCart
+                        user.cart.productList = newCart
                     }
                     productFound = true 
                 }
@@ -254,6 +321,7 @@ exports.removeFromCart = (req, res, next) => {
                 err.message="Product is not in the cart!"
                 throw err
             }
+            user.cart.total = user.cart.total - totalDiffrence
             return user.save()
         }
         else{
@@ -273,6 +341,12 @@ exports.removeFromCart = (req, res, next) => {
 };
 
 exports.getCart = (req, res, next) => {
+    if(req.userId != req.body._id){
+        let err = new Error()
+        err.statusCode=422
+        err.message="Invalid Token!"
+        throw err
+    }
     const _id = req.body._id
     User.findById(_id).populate('cart.product')
     .then(result =>{
@@ -293,42 +367,79 @@ exports.getCart = (req, res, next) => {
     })
 };
 
+exports.adminGetCart = (req, res, next) => {
+    if(isAdmin){
+        const _id = req.body._id
+        User.findById(_id).populate('cart.product')
+        .then(result =>{
+            if(result != null){
+                res.status(200).json({ message: "Cart found!", cart: result.cart});
+            }
+            else{
+                let err = new Error()
+                err.statusCode=404
+                err.message="User not found!"
+                throw err
+            }
+        }).catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+              }
+              next(err);
+        })
+    }
+    else{
+        let err = new Error()
+        err.statusCode=422
+        err.message="Not authenticated!"
+        next(err)
+    }
+};
+
 exports.addAdmin = (req, res, next) => {
-    const name = req.body.name
-    const username = req.body.username
-    const email = req.body.email
-    const imageUrl = req.body.imageUrl || ""
-    let user 
-    User.findOne({'email': email})
-    .then(result =>{
-        if(result != null){
-            let err = new Error()
-            err.statusCode=422
-            err.message="Email already exist!"
-            throw err
-        }
-        else{
-            return User.findOne({'username': username})
-        }
-    }).then(result => {
-        if(result != null){
-            let err = new Error()
-            err.statusCode=422
-            err.message="Username already exist!"
-            throw err
-        }
-        else{
-            return bcrypt.hash(req.body.password , 12)
-        }
-    }).then(result =>{
-        user = new User({name: name, username: username, email: email, password: result, imageUrl: imageUrl, isAdmin: true})
-        return user.save()
-    }).then(result => {
-        res.status(201).json({ message: "User created!", user: user});
-    }).catch(err => {
-        if (!err.statusCode) {
-            err.statusCode = 500;
-          }
-          next(err);
-    })
+    if(isAdmin){
+        const name = req.body.name
+        const username = req.body.username
+        const email = req.body.email
+        const imageUrl = req.body.imageUrl || ""
+        let user 
+        User.findOne({'email': email})
+        .then(result =>{
+            if(result != null){
+                let err = new Error()
+                err.statusCode=422
+                err.message="Email already exist!"
+                throw err
+            }
+            else{
+                return User.findOne({'username': username})
+            }
+        }).then(result => {
+            if(result != null){
+                let err = new Error()
+                err.statusCode=422
+                err.message="Username already exist!"
+                throw err
+            }
+            else{
+                return bcrypt.hash(req.body.password , 12)
+            }
+        }).then(result =>{
+            user = new User({name: name, username: username, email: email, password: result, imageUrl: imageUrl, isAdmin: true})
+            return user.save()
+        }).then(result => {
+            res.status(201).json({ message: "User created!", user: user});
+        }).catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+              }
+              next(err);
+        })
+    }
+    else{
+        let err = new Error()
+        err.statusCode=422
+        err.message="Not authenticated!"
+        next(err)
+    }
 };
